+++ R/revdepcheck.R
@@ -40,6 +40,8 @@
 #'   complete. Default is 10 minutes.
 #' @param num_workers Number of parallel workers to use
 #' @param bioc Also check revdeps that live in Bioconductor?
+#' @param cran Should cran mirror be attached to getOpion("repos") if it
+#' is not already present.
 #' @param env Environment variables to set for the install and check
 #'   processes. See [revdep_env_vars()].
 #'
@@ -58,6 +60,7 @@ revdep_check <- function(pkg = ".",
                          timeout = as.difftime(10, units = "mins"),
                          num_workers = 1,
                          bioc = TRUE,
+                         cran = TRUE,
                          env = revdep_env_vars()) {
 
   pkg <- pkg_check(pkg)
@@ -70,8 +73,10 @@ revdep_check <- function(pkg = ".",
   repeat {
     stage <- db_metadata_get(pkg, "todo") %|0|% "init"
     switch(stage,
-      init =    revdep_init(pkg, dependencies = dependencies, bioc = bioc),
-      install = revdep_install(pkg, quiet = quiet, env = env),
+      init =    revdep_init(pkg, dependencies = dependencies, 
+                            bioc = bioc, cran = cran),
+      install = revdep_install(pkg, quiet = quiet, env = env, 
+                               bioc = bioc, cran = cran),
       run =     revdep_run(pkg, quiet = quiet, timeout = timeout,
                            num_workers = num_workers, env = env),
       report =  revdep_final_report(pkg),
@@ -103,7 +108,7 @@ revdep_setup <- function(pkg = ".") {
 revdep_init <- function(pkg = ".",
                          dependencies = c("Depends", "Imports",
                                           "Suggests", "LinkingTo"),
-                         bioc = TRUE) {
+                         bioc = TRUE, cran = TRUE) {
 
   pkg <- pkg_check(pkg)
   pkgname <- pkg_name(pkg)
@@ -111,7 +116,7 @@ revdep_init <- function(pkg = ".",
 
   "!DEBUG getting reverse dependencies for `basename(pkg)`"
   status("INIT", "Computing revdeps")
-  revdeps <- cran_revdeps(pkgname, dependencies, bioc = bioc)
+  revdeps <- cran_revdeps(pkgname, dependencies, bioc = bioc, cran = cran)
   db_todo_add(pkg, revdeps)
 
   db_metadata_set(pkg, "todo", "install")
@@ -121,7 +126,8 @@ revdep_init <- function(pkg = ".",
   invisible()
 }
 
-revdep_install <- function(pkg = ".", quiet = FALSE, env = character()) {
+revdep_install <- function(pkg = ".", quiet = FALSE, env = character(), 
+                           bioc = bioc, cran = TRUE) {
   pkg <- pkg_check(pkg)
   pkgname <- pkg_name(pkg)
 
@@ -147,7 +153,7 @@ revdep_install <- function(pkg = ".", quiet = FALSE, env = character()) {
       dir_find(pkg, "old"),
       rlang::with_options(
         warn = if (fail_on_warn) 2 else 1,
-        install_packages(pkgname, quiet = quiet, repos = get_repos(bioc = TRUE), upgrade = "always")
+        install_packages(pkgname, quiet = quiet, repos = get_repos(bioc = bioc, cran = cran), upgrade = "always")
       )
     )
   )
@@ -161,7 +167,7 @@ revdep_install <- function(pkg = ".", quiet = FALSE, env = character()) {
       dir_find(pkg, "new"),
       rlang::with_options(
         warn = if (fail_on_warn) 2 else 1,
-        install_local(pkg, quiet = quiet, repos = get_repos(bioc = TRUE), force = TRUE, upgrade = "always")
+        install_local(pkg, quiet = quiet, repos = get_repos(bioc = bioc, cran = cran), force = TRUE, upgrade = "always")
       )
     )
   )