+++ NEWS.md
@@ -1,5 +1,7 @@
-
 # 1.0.0.9000
+
+* `revdep_report()` now saves the results of `revdep_report_cran()` to
+  `revdep/cran.md` (#204)
 
 * Exported `revdep_report()` to write the current results to README.md
   and problems.md.